+++ src/ContaoManager/Plugin.php
@@ -1,15 +1,5 @@
 <?php
 
-/**
- * Contao Bootstrap Layout.
- *
- * @package    contao-bootstrap
- * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2014-2017 netzmacht creative David Molineus
- * @license    LGPL 3.0
- * @filesource
- */
-
 declare(strict_types=1);
 
 namespace ContaoBootstrap\Layout\ContaoManager;
@@ -21,17 +11,12 @@ use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
 use ContaoBootstrap\Core\ContaoBootstrapCoreBundle;
 use ContaoBootstrap\Layout\ContaoBootstrapLayoutBundle;
 
-/**
- * Contao Manager plugin.
- *
- * @package ContaoBootstrap\Layout\ContaoManager
- */
-class Plugin implements BundlePluginInterface
+final class Plugin implements BundlePluginInterface
 {
     /**
      * {@inheritdoc}
      */
-    public function getBundles(ParserInterface $parser)
+    public function getBundles(ParserInterface $parser): array
     {
         $bundleConfig = BundleConfig::create(ContaoBootstrapLayoutBundle::class)
             ->setLoadAfter(