+++ src/DependencyInjection/ContaoBootstrapLayoutExtension.php
@@ -1,15 +1,5 @@
 <?php
 
-/**
- * Contao Bootstrap Layout.
- *
- * @package    contao-bootstrap
- * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2014-2017 netzmacht creative David Molineus
- * @license    LGPL 3.0
- * @filesource
- */
-
 declare(strict_types=1);
 
 namespace ContaoBootstrap\Layout\DependencyInjection;
@@ -19,20 +9,15 @@ use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\Extension;
 use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
 
-/**
- * Class ContaoBootstrapLayoutExtension.
- *
- * @package ContaoBootstrap\Layout\DependencyInjection
- */
-class ContaoBootstrapLayoutExtension extends Extension
+final class ContaoBootstrapLayoutExtension extends Extension
 {
     /**
      * {@inheritdoc}
      */
-    public function load(array $configs, ContainerBuilder $container)
+    public function load(array $configs, ContainerBuilder $container): void
     {
         $loader = new YamlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
-        $loader->load('services.yml');
-        $loader->load('listener.yml');
+        $loader->load('services.yaml');
+        $loader->load('listener.yaml');
     }
 }