+++ src/Listener/Dca/LayoutDcaListener.php
@@ -1,52 +1,46 @@
 <?php
 
-/**
- * Contao Bootstrap Layout.
- *
- * @package    contao-bootstrap
- * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2014-2017 netzmacht creative David Molineus
- * @license    LGPL 3.0
- * @filesource
- */
-
 declare(strict_types=1);
 
 namespace ContaoBootstrap\Layout\Listener\Dca;
 
+use Contao\CoreBundle\ServiceAnnotation\Callback;
 use Contao\DataContainer;
+use Contao\StringUtil;
 use ContaoBootstrap\Core\Config;
 use Doctrine\DBAL\Connection;
+use Netzmacht\Contao\Toolkit\Dca\DcaManager;
+use Netzmacht\Contao\Toolkit\Dca\Listener\AbstractListener;
 
-/**
- * Dca Helper class for tl_layout.
- *
- * @package Netzmacht\Bootstrap\DataContainer
- */
-final class LayoutDcaListener
+use function array_filter;
+use function array_keys;
+use function array_values;
+use function serialize;
+
+final class LayoutDcaListener extends AbstractListener
 {
+    /** @var string */
+    // phpcs:ignore SlevomatCodingStandard.TypeHints.PropertyTypeHint.MissingNativeTypeHint
+    protected static $name = 'tl_layout';
+
     /**
      * Bootstrap config.
-     *
-     * @var Config
      */
-    private $config;
+    private Config $config;
 
     /**
      * Database connection.
-     *
-     * @var Connection
      */
-    private $connection;
+    private Connection $connection;
 
     /**
-     * Layout constructor.
-     *
      * @param Config     $config     Bootstrap config.
      * @param Connection $connection Database connection.
      */
-    public function __construct(Config $config, Connection $connection)
+    public function __construct(Config $config, Connection $connection, DcaManager $dcaManager)
     {
+        parent::__construct($dcaManager);
+
         $this->config     = $config;
         $this->connection = $connection;
     }
@@ -54,8 +48,7 @@ final class LayoutDcaListener
     /**
      * Set the default viewport.
      *
-     * @return void
-     *
+     * @Callback(table="tl_layout", target="config.onload")
      * @SuppressWarnings(PHPMD.Superglobals)
      */
     public function setDefaultViewPort(): void
@@ -66,18 +59,52 @@ final class LayoutDcaListener
     /**
      * Disable contao framework.
      *
-     * @param DataContainer $dataContainer Data container driver.
-     *
-     * @return void
+     * @Callback(table="tl_layout", target="config.onsubmit")
      */
     public function disableFramework(DataContainer $dataContainer): void
     {
-        if ($dataContainer->activeRecord->layoutType === 'bootstrap' && $dataContainer->activeRecord->framework) {
-            $dataContainer->activeRecord->framework = [];
-
-            $statement = $this->connection->prepare('UPDATE tl_layout SET framework = \'\' WHERE id=?');
-            $statement->bindValue(1, $dataContainer->id);
-            $statement->execute();
+        if (! $dataContainer->activeRecord) {
+            return;
         }
+
+        if ($dataContainer->activeRecord->layoutType !== 'bootstrap' || ! $dataContainer->activeRecord->framework) {
+            return;
+        }
+
+        /** @psalm-var array<string,bool> $supportedFrameworkCss */
+        $supportedFrameworkCss = $this->config->get('layout.contao_framework_css', []);
+
+        $dataContainer->activeRecord->framework = array_values(
+            array_filter(
+                StringUtil::deserialize($dataContainer->activeRecord->framework, true),
+                static fn (string $value) => $supportedFrameworkCss[$value] ?? true,
+            )
+        );
+
+        $this->connection->executeStatement(
+            'UPDATE tl_layout SET framework = :framework WHERE id= :id',
+            [
+                'framework' => serialize($dataContainer->activeRecord->framework),
+                'id' => $dataContainer->id,
+            ]
+        );
+    }
+
+    /**
+     * @Callback(table="tl_layout", target="fields.framework.load")
+     */
+    public function frameworkOptions(string $value, DataContainer $dataContainer): string
+    {
+        if (! $dataContainer->activeRecord || $dataContainer->activeRecord->layoutType !== 'bootstrap') {
+            return $value;
+        }
+
+        /** @psalm-var array<string,bool> $supportedFrameworkCss */
+        $supportedFrameworkCss = $this->config->get('layout.contao_framework_css', []);
+        $files                 = array_keys(array_filter($supportedFrameworkCss));
+
+        $this->getDefinition()->set(['fields', 'framework', 'options'], $files);
+
+        return $value;
     }
 }