+++ src/View/Template/AbstractFilter.php
@@ -1,44 +1,29 @@
 <?php
 
-/**
- * Contao Bootstrap
- *
- * @package    contao-bootstrap
- * @subpackage Layout
- * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2017 netzmacht David Molineus. All rights reserved.
- * @license    LGPL 3.0
- * @filesource
- */
+declare(strict_types=1);
 
 namespace ContaoBootstrap\Layout\View\Template;
 
 use ContaoBootstrap\Core\Environment;
 
-/**
- * Class AbstractFilter
- *
- * @package ContaoBootstrap\Layout\View\Template
- */
+use function is_array;
+use function strcasecmp;
+use function strlen;
+use function substr;
+
 abstract class AbstractFilter
 {
     /**
      * Bootstrap environment.
-     *
-     * @var Environment
      */
-    private $environment;
+    private Environment $environment;
 
     /**
      * The config key for the templates.
-     *
-     * @var string
      */
-    private $templateConfigKey;
+    private string $templateConfigKey;
 
     /**
-     * AbstractFilter constructor.
-     *
      * @param Environment $environment       The bootstrap environment.
      * @param string      $templateConfigKey The templates config key.
      */
@@ -50,8 +35,6 @@ abstract class AbstractFilter
 
     /**
      * Get environment.
-     *
-     * @return Environment
      */
     protected function getEnvironment(): Environment
     {
@@ -62,14 +45,12 @@ abstract class AbstractFilter
      * Check if template name is supported.
      *
      * @param string $templateName Template name.
-     *
-     * @return bool
      */
     protected function isTemplateNameSupported(string $templateName): bool
     {
         $templateNames = $this->getEnvironment()->getConfig()->get($this->templateConfigKey);
 
-        if (!is_array($templateNames)) {
+        if (! is_array($templateNames)) {
             return false;
         }
 
@@ -78,8 +59,9 @@ abstract class AbstractFilter
                 return true;
             }
 
-            if (substr($supported, -1) === '*'
-                && 0 == strcasecmp(substr($supported, 0, -1), substr($templateName, 0, (strlen($supported) - 1)))
+            if (
+                substr($supported, -1) === '*'
+                && strcasecmp(substr($supported, 0, -1), substr($templateName, 0, strlen($supported) - 1)) === 0
             ) {
                 return true;
             }